NonRecPairs Initial Feedback Incorporation
More trees added to the test file (nonRecPairs.html) along with explanations and predictions. Testing moved out of console and into web-page.

Clearer comments added to nonrecursivity.js regarding evaluation of pairs that do not involve immidiate domination.

A slight issue arose testing a tree where a phi dominates an iota. It should be a quick fix, but I want to commit this code in case I break it.