nonRecParent addition
Added a function for the constraint Non-Recursivity by Parent: "assign one violation for every node of category c that immidiately dominates at least one node of category c."

The function nonRecParent was added to the bottom of constraints/nonrecursivity.js. A test was added as the file test/nonRecParent.html. Everything appears to be working well, further testing might be necessary, but I can't think of anything that might be an issue.