Added syll to Strong Start Elfner in Interface
Added a checkbox for syllables in StrongStart for the interface (line 322). While the checkbox causes a constraint to be added to to the output tableau, I was unable to get any candidates to violate this constraint. In the JS console I was able to manually produce a ptree which successfuly incurred a violation of strongStartElfner with the cat "syll", which leads me to believe that the reason I did not get any violations from the interface was that I was unable to get GEN to produce a candidate with syllables.