generalized binarity constraints added, issues with iota
added three constraints: binMaxLeaves, binMaxLeavesGradient and binMinLeaves. Added test file for these constraints.

Tests look good when non-binary phis are targeted, but there are some issues with iota. When non-binary i is targeted, getDescendentsOfCat fails to return the correct number of phis. this leads to binMaxLeaves constraints to only return 0 and binMinLeaves constraint to only return 1.  I was able to get getDescendentsOfCat to return 1 phi for the tree {(word)} using the console, but in practice it only returns 0 phis.

Using the console, I was able to get promising results for non binary w.