fixed getDescendantsOfCat function
getDescendantsofCat was missing the lines of code which pushed nodes to the output array "descendants" when they matched the category specified. originally, this was suplimented by an else if statement which pushed terminal nodes to descendants. this statement worked fine for prosodic words since they are usually terminal, but when phis needed to be counted, they were usually not terminal and were therfore never pushed to descendants. once lines 159-161 were added, all nodes of the specified category were pushed to descendants, so the else if statement double counted terminals. for this reason i commeneted the block out, leaving an explanation in the code. fixing this helper function fixed issues with bin--Leaves-i mentioned in commit 9c4b681ff1efd604ce784f7127a82ddf351eaf15.