Strated matchMaxSP constraint
Created a function and test file for matchMaxSP constraint. So far unable to get any violations, though.

Uses markMinMax function from recursiveCatEvals, which has a slight issue. The function never marks the root of a tree as maximal, since the program doesn't know that the root has no parent node. Not quite sure what to do about that.

Also I added a semicolon to the end of line 53 of recursivCatEvals.js