matchMaxPS bug bodged
The constraint is working, but the solution is probably not ideal in the long run.

I added a function called clearMinMax to recursiveCatEvals.js which sets the values of isMin and isMax to "undefined" (void(0)) for an entire tree.

I added a helper function for matchMaxSP so that markMinMax and clearMinMax can be run only once while the helper function runs recursively.

This sidesteps the problem of re-used nodes but any future constraints that rely on markMinMax for prosodic trees will have to use helper functions to first call markMinMax, then call the helper function, which must be recursive, then call clearMinMax.

Maybe this functionality should be built into markMinMax, but for now this fixes the issue I have been dealing with.