markMinMax long-term solution
integrating solution for markMinMax so that it rewrites isMin and isMax properties when appropriate. Testing in progress.

In a recursive function, markMinMax will be called on a tree and all of its subtrees and will not always be called with the parCat argument. In this case, all nodes wil be overwritten with isMax : true even when a node is not maximal.

To circumvent this, I added the property parentCat to the tree so that if a subtree is reused by GEN, the function will catch any outdated isMin/isMax values and reset them.