Constraint Definition Misinterpretation
Up until this point I had interpreted the constraint as only only the syntactic node must be maximal for MatchMaxSP and only the prosodic node must be maximal in MatchMaxPS. On closer inspection, the definition in the issue for this constraint specifies that both a node and its matching node must be maximal for a violation to be assigned.

This commit fixes the constraint so that it follows the intended definition and preserves my faulty interpretation as matchMaxSyntax and matchMaxProsody.

The fix involved defining the helper function hasMaxMatch, which is basically hasMatch but only returns true of the matching node is maximal.

When given 3 terminals, the outputs of GEN return the exact same number of violations for matchMaxSP and MatchMaxPS as MatchMaxSyntax and MatchMax Prosody respectively.