updating markMinMax so it's harder to break
the last update to markMinMax made it too complecated, so this update makes it simple again.

This time it is harder to break when using in a recursive function and, after simplifying things, I noticed that my old version had a bug which was fixed in this version.

I also updated the test file to reflect what it was actually testing instead of saying that it was testing nonRecPairs