Tests Complete and Bugs Fixed
Added a function to equalStrength.js to return a syntactic category given a prosodic category based on categoryPairings in prosodicHierarchy.js.
Made two test files, one for each edge, since 6 constraints is a little hard to read. Tested a random sample of 12.5% of the candidates and all constraints performed exactly as expected.