addTones now takes a string argument
the GEN option addTones now takes a string as its value instead of a boolean. The string should be the name of a function that adds tones to a tree and that function will be called. If the user sets the value of addtones to true, addJapaneseTones will still be called and a log message will alert the user to the changes that have been made. Otherwise, exceptions in calling the specified function result in an error message explaining the requirements of the option.