added rootCategory and terminalCategory to GEN options
Created options in GEN for rootCategory and terminalCategory. If these are not specified, they are set to defaults based on recursiveCategory.
If rootCategory, recursiveCategory or terminalCategory are specified as categories that do not exist in pCat, that category is added to pCat. I guessed where they would be most useful.
The number of candidates is doubled with addPhiWrapped only if rootCategory and recursiveCategory are different. This avoids structures like {i {i (...) (...)}}  which we probably don't want.
If you specify a category that doesn't exist, the GEN option obeysNonRecursivity doesn't quite work (see the final tableau in Recursive_GEN_options.html)