fixed obeysHeadedness
instead of checking just the root node for headedness before adding it to the candidates, GEN now checks the whole tree with the function obeysHeadedness.

test file confirms that everything is working properly. All of the trees generated when obeysHeadedness: true obey headedness and some extra code shows that all of the trees that obey headedness are added.(generated some tableaux with only trees that violate headedness. All of these trees do violate headedness and the number of trees generated plus the number of trees generted when obeysHeadedness:true equals the number of trees generated when obeysHeadedness:false.)