Clear All now clears the file upload
Clear All button now clears the file upload. Now, when the user hits "clear all", "upload file" says "no file selected". This is desirable because whatever analysis was chosen by the user will have been cleared out. This should not be called in clearAnalysis() becasue clearAnalysis() is called when the file is selected