a quick fix and some conOptions updates
A quick fix: it turns out that the typeOf for a NodeList does not return Array, so I had to change a line of code I wrote in my last commit.

Some conOptions: as it stants, the infrastucture to deal with constraint options assumes that there are multiple options and treats the options list as an array. If there is only one option, you can't treat it as an array, so I made checks to avoid errors.