define & test strongStartDeep
This implements what i believe was Nick Kalivoda's intention for the constraint strongStartInitial. I've renamed it because "start" and  "initial" are redundant. The distinctive characteristic of this constraint is that it penalizes a weak start to the specified category at any level, not that it looks at the beginning of the sentence.